garbage_front_end_rt_v4/requirements.txt
alembic==1.3.2
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
torch==1.4.0+cu92
torchvision==0.5.0+cu92
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
